refactoring - present the health aspect as part of the interface
