+++ shutdown/health_test.go
@@ -1,19 +1,24 @@
-package health
+package shutdown
 
 import (
+	"github.com/ThomasObenaus/go-base/shutdown/stop"
+	"github.com/rs/zerolog"
 	"github.com/stretchr/testify/assert"
 	"testing"
 )
 
 func Test_is_healthy_when_started(t *testing.T) {
-	handler := Health{}
+	handler := ShutdownHandler{}
 	err := handler.IsHealthy()
 
 	assert.NoError(t, err)
 }
 
 func Test_is_unhealthy_when_shutting_down(t *testing.T) {
-	handler := Health{}
+	handler := ShutdownHandler{
+		stoppableItems: &stop.OrderedStoppableList{},
+		logger:         zerolog.Nop(),
+	}
 	handler.ShutdownSignalReceived()
 	err := handler.IsHealthy()
 
@@ -21,7 +26,10 @@ func Test_is_unhealthy_when_shutting_down(t *testing.T) {
 }
 
 func Test_reports_health_status_depending_on_state(t *testing.T) {
-	handler := Health{}
+	handler := ShutdownHandler{
+		stoppableItems: &stop.OrderedStoppableList{},
+		logger:         zerolog.Nop(),
+	}
 
 	status := handler.String()
 	assert.Equal(t, "ShutdownHandler (shutdown in progress=false)", status)
