+++ shutdown/legacy_test.go
@@ -3,7 +3,7 @@ package shutdown
 import (
 	"fmt"
 	"github.com/golang/mock/gomock"
-	"github.com/stretchr/testify/assert"
+	"github.com/rs/zerolog"
 	"testing"
 )
 
@@ -32,17 +32,14 @@ func Test_logs_failure_if_stoppable_can_not_be_added_in_front(t *testing.T) {
 	defer mockCtrl.Finish()
 
 	mockStop := NewMockstopIF(mockCtrl)
-	mockLog := NewMocklogIF(mockCtrl)
 	stoppable := NewMockStoppable(mockCtrl)
 	shutdownHandler := ShutdownHandler{
 		stoppableItems: mockStop,
-		log:            mockLog,
 	}
 
 	// EXPECT
 	mockStop.EXPECT().AddToFront(stoppable).Return(fmt.Errorf("some error"))
 	stoppable.EXPECT().String().Return("some service")
-	mockLog.EXPECT().LogCanNotAddService("some service")
 
 	// WHEN
 	shutdownHandler.Register(stoppable, true)
@@ -70,17 +67,14 @@ func Test_logs_failure_if_stoppable_can_not_be_added_to_back(t *testing.T) {
 	defer mockCtrl.Finish()
 
 	mockStop := NewMockstopIF(mockCtrl)
-	mockLog := NewMocklogIF(mockCtrl)
 	stoppable := NewMockStoppable(mockCtrl)
 	shutdownHandler := ShutdownHandler{
 		stoppableItems: mockStop,
-		log:            mockLog,
 	}
 
 	// EXPECT
 	mockStop.EXPECT().AddToFront(stoppable).Return(fmt.Errorf("some error"))
 	stoppable.EXPECT().String().Return("some service")
-	mockLog.EXPECT().LogCanNotAddService("some service")
 
 	// WHEN
 	shutdownHandler.Register(stoppable)
@@ -123,79 +117,22 @@ func Test_logs_all_stop_related_events(t *testing.T) {
 	mockCtrl := gomock.NewController(t)
 	defer mockCtrl.Finish()
 
-	mockLog := NewMocklogIF(mockCtrl)
 	mockStop := NewMockstopIF(mockCtrl)
-	mockHealth := NewMockhealthIF(mockCtrl)
+	logger := zerolog.Nop()
 	shutdownHandler := ShutdownHandler{
 		stoppableItems: mockStop,
-		log:            mockLog,
-		health:         mockHealth,
+		logger:         logger,
 	}
 
 	// IGNORE
-	mockHealth.EXPECT().ShutdownSignalReceived().AnyTimes()
 
 	// EXPECT
 	gomock.InOrder(
-		mockLog.EXPECT().ShutdownSignalReceived(),
-		mockStop.EXPECT().StopAllInOrder(mockLog),
+		mockStop.EXPECT().StopAllInOrder(logger),
 	)
 
 	// WHEN
 	shutdownHandler.ShutdownSignalReceived()
-}
-
-func Test_notifies_health_monitor_on_service_stop(t *testing.T) {
-	// GIVEN
-	mockCtrl := gomock.NewController(t)
-	defer mockCtrl.Finish()
-
-	mockLog := NewMocklogIF(mockCtrl)
-	mockStop := NewMockstopIF(mockCtrl)
-	mockHealth := NewMockhealthIF(mockCtrl)
-	shutdownHandler := ShutdownHandler{
-		stoppableItems: mockStop,
-		log:            mockLog,
-		health:         mockHealth,
-	}
-
-	// IGNORE
-	mockLog.EXPECT().ShutdownSignalReceived().AnyTimes()
-	mockStop.EXPECT().StopAllInOrder(gomock.Any()).AnyTimes()
-
-	// EXPECT
-	mockHealth.EXPECT().ShutdownSignalReceived()
-
-	// WHEN
-	shutdownHandler.ShutdownSignalReceived()
-}
-
-func Test_uses_health_monitor_to_report_health_status(t *testing.T) {
-	// GIVEN
-	mockCtrl := gomock.NewController(t)
-	defer mockCtrl.Finish()
-
-	mockHealth := NewMockhealthIF(mockCtrl)
-	shutdownHandler := ShutdownHandler{
-		health: mockHealth,
-	}
-
-	// EXPECT
-	gomock.InOrder(
-		mockHealth.EXPECT().IsHealthy().Return(nil),
-		mockHealth.EXPECT().IsHealthy().Return(fmt.Errorf("some error")),
-		mockHealth.EXPECT().String().Return("some status"),
-	)
-
-	// WHEN
-	err := shutdownHandler.IsHealthy()
-	assert.NoError(t, err)
-
-	err = shutdownHandler.IsHealthy()
-	assert.Error(t, err)
-
-	status := shutdownHandler.String()
-	assert.Equal(t, "some status", status)
 }
 
 func Test_informs_stop_that_it_should_stop(t *testing.T) {
@@ -203,21 +140,17 @@ func Test_informs_stop_that_it_should_stop(t *testing.T) {
 	mockCtrl := gomock.NewController(t)
 	defer mockCtrl.Finish()
 
-	mockLog := NewMocklogIF(mockCtrl)
+	logger := zerolog.Nop()
 	mockStop := NewMockstopIF(mockCtrl)
-	mockHealth := NewMockhealthIF(mockCtrl)
 	shutdownHandler := ShutdownHandler{
 		stoppableItems: mockStop,
-		log:            mockLog,
-		health:         mockHealth,
+		logger:         logger,
 	}
 
 	// IGNORE
-	mockHealth.EXPECT().ShutdownSignalReceived().AnyTimes()
-	mockLog.EXPECT().ShutdownSignalReceived()
 
 	// EXPECT
-	mockStop.EXPECT().StopAllInOrder(mockLog)
+	mockStop.EXPECT().StopAllInOrder(logger)
 
 	// WHEN
 	shutdownHandler.ShutdownSignalReceived()
