+++ shutdown/mock_interfaces_test.go
@@ -121,66 +121,3 @@ func (mr *MocksignalHandlerIFMockRecorder) WaitForSignal() *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "WaitForSignal", reflect.TypeOf((*MocksignalHandlerIF)(nil).WaitForSignal))
 }
-
-// MockhealthIF is a mock of healthIF interface.
-type MockhealthIF struct {
-	ctrl     *gomock.Controller
-	recorder *MockhealthIFMockRecorder
-}
-
-// MockhealthIFMockRecorder is the mock recorder for MockhealthIF.
-type MockhealthIFMockRecorder struct {
-	mock *MockhealthIF
-}
-
-// NewMockhealthIF creates a new mock instance.
-func NewMockhealthIF(ctrl *gomock.Controller) *MockhealthIF {
-	mock := &MockhealthIF{ctrl: ctrl}
-	mock.recorder = &MockhealthIFMockRecorder{mock}
-	return mock
-}
-
-// EXPECT returns an object that allows the caller to indicate expected use.
-func (m *MockhealthIF) EXPECT() *MockhealthIFMockRecorder {
-	return m.recorder
-}
-
-// IsHealthy mocks base method.
-func (m *MockhealthIF) IsHealthy() error {
-	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "IsHealthy")
-	ret0, _ := ret[0].(error)
-	return ret0
-}
-
-// IsHealthy indicates an expected call of IsHealthy.
-func (mr *MockhealthIFMockRecorder) IsHealthy() *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "IsHealthy", reflect.TypeOf((*MockhealthIF)(nil).IsHealthy))
-}
-
-// ShutdownSignalReceived mocks base method.
-func (m *MockhealthIF) ShutdownSignalReceived() {
-	m.ctrl.T.Helper()
-	m.ctrl.Call(m, "ShutdownSignalReceived")
-}
-
-// ShutdownSignalReceived indicates an expected call of ShutdownSignalReceived.
-func (mr *MockhealthIFMockRecorder) ShutdownSignalReceived() *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ShutdownSignalReceived", reflect.TypeOf((*MockhealthIF)(nil).ShutdownSignalReceived))
-}
-
-// String mocks base method.
-func (m *MockhealthIF) String() string {
-	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "String")
-	ret0, _ := ret[0].(string)
-	return ret0
-}
-
-// String indicates an expected call of String.
-func (mr *MockhealthIFMockRecorder) String() *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "String", reflect.TypeOf((*MockhealthIF)(nil).String))
-}
